deleteTeam fixes
